Change the default for apply_rng in hk.transform.
This is a breaking change that changes the default for `apply_rng` from `False` to `True`.

Too many signatures are making the library harder to use and reason about. Suddenly having to add a seed is also very brittle. Thus, the default behavior is now to assume an RNG is used.

For equivalent behavior please either use:

1) `f = hk.without_apply_rng(hk.transform(f))` and call `f.apply(params, ...)`.
2) or `f = hk.transform(f)` and pass a `None` rng to your apply function: `f.apply(params, None, ...)`.

PiperOrigin-RevId: 319776685
Change-Id: I9e591a4747a5e829a3e522e957fc9d5b5ea5b3ee